Manufacturer Controller

+++ MedSave/Model/Manufacturer/ContactManufacturer.cs
@@ -4,5 +4,5 @@ public class ContactManufacturer
 {
     public long ContactManuId { get; set; }
     public string EmailManu { get; set; }
-    public int PhoneNumberManu { get; set; }
+    public long PhoneNumberManu { get; set; }
 }
