+++ MedSave/Model/Manufacturer/Manufacturer.cs
@@ -4,7 +4,7 @@ public class Manufacturer
 {
     public long ManufacId { get; set; }
     public string NameManu { get; set; }
-    public int Cnpj { get; set; }
+    public long Cnpj { get; set; }
     
     public long ContactManuId { get; set; }
     public ContactManufacturer ContactManufacturer { get; set; }
